// src/config.d.ts
import "dotenv/config";
export declare const config: {
    discordToken: string;
    discordClientId: string;
    discordGuildId: string;
    supabaseUrl: string;
    supabaseSecretKey: string;
    openRouterApiKey: string | undefined;
    openRouterEvaluationModel: string;
    e2bApiKey: string | undefined;
    e2bSandboxTemplate: string;
};
//# sourceMappingURL=config.d.ts.map